Rename and move tests for clarity
